+++ OpenBurnBarDaemon/Sources/OpenBurnBarDaemon/OpenBurnBarDaemonConfiguration.swift
@@ -89,7 +89,7 @@ public enum BurnBarDaemonPaths {
 }
 
 public enum BurnBarDaemonVersion {
-    public static let current = "1.0.14"
+    public static let current = "1.0.15"
 }
 
 public struct BurnBarGatewayConfiguration: Codable, Hashable, Sendable {